extract more code from monolith
